Move JMH content to jmh package
